% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{addProgenyData}
\alias{addProgenyData}
\title{Add progeny information to data after cross, doubledHaploid, or selfFertilize}
\usage{
addProgenyData(bsl, geno, pedigree)
}
\arguments{
\item{bsl}{the list that has all the objects for one simulation}

\item{geno}{the genotypes of the progeny}

\item{pedigree}{the three-column pedigree of the progeny (last col: DH, outbred, self)}
}
\value{
data with progeny information added
}
\description{
Add progeny information to data after cross, doubledHaploid, or selfFertilize
}

