% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculationFunc.R
\name{calcPhenotypicValue}
\alias{calcPhenotypicValue}
\title{calcPhenotypicValue}
\usage{
calcPhenotypicValue(gv, nRep, errorVar)
}
\arguments{
\item{gv}{genotypic values}

\item{nRep}{how many replications of phenotypic values: used for multiple locations and/or years}

\item{errorVar}{error variance}
}
\description{
calcPhenotypicValue
}

