% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineCosts.R
\name{defineCosts}
\alias{defineCosts}
\title{Define the costs that go into breeding
Default for some costs is zero because they probably belong to fixed costs}
\usage{
defineCosts(sEnv = simEnv, phenoCost = NULL, genoCost = 0.25,
  crossCost = 1, selfCost = 1, doubHapCost = 5, predCost = 0,
  selectCost = 0, locCost = 0, yearCost = 0)
}
\arguments{
\item{sEnv}{the environment that BSL functions operate in. Default is "simEnv" so use that to avoid specifying when calling functions}

\item{phenoCost}{named vector: names are the plot types and contents are the cost for one plot of that type (default: Standard=1; if plotTypes have been defined in defineVariances and phenoCost is not specified, all plotTypes are given a cost of 1)}

\item{genoCost}{scalar: cost to genotype one individual default (0.25)}

\item{crossCost}{scalar: cost of creating a new individual from a cross (1)}

\item{selfCost}{scalar: the cost of creating a selfed seed (1)}

\item{doubHapCost}{scalar: the cost of creating a doubled haploid seed (5)}

\item{predCost}{scalar: the cost of running the analysis to make predictions (0)}

\item{selectCost}{scalar: the cost of running the analysis to do selection (0)}

\item{locCost}{scalar: the cost of maintaining a location for a year (0)}

\item{yearCost}{scalar: the cost of program upkeep for a year (0)}
}
\value{
Breeding scheme simulation object supplemented with a list of costs
}
\description{
Define the costs that go into breeding
Default for some costs is zero because they probably belong to fixed costs
}

