% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.R
\name{genotype}
\alias{genotype}
\title{Genotype markers}
\usage{
genotype(sEnv = simEnv, popID = NULL)
}
\arguments{
\item{sEnv}{the environment that BSL functions operate in. Default is "simEnv" so use that to avoid specifying when calling functions}

\item{popID}{population ID to be genotyped (default: all populations)}
}
\value{
marker genotype and the all information created before (list)
}
\description{
Genotype markers
}

