% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputResults.R
\name{outputResults}
\alias{outputResults}
\title{Save the results}
\usage{
outputResults(sEnv = simEnv, summarize = T, directory = ".",
  saveDataFileName = "BSLoutput")
}
\arguments{
\item{sEnv}{the environment that BSL functions operate in. Default is "simEnv" so use that to avoid specifying when calling functions}

\item{summarize}{if T a result averaged over all the replications is saved, if F each replication's result is saved}

\item{directory}{the directory to which the output will be saved (Enclose the name in double quotation!) (default: the current directory)}

\item{saveDataFileName}{the file name to save the simulated data with double-quotation, like "result1_1". (default: "BSLoutput")}
}
\value{
The simulation results (The output data was saved as BSLoutput.RData. After you load the data in R, you can find the data named as BSLoutput.)
}
\description{
Save the results
}

