% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{phasedHapMap2mat}
\alias{phasedHapMap2mat}
\title{Transform a data.frame with a hapmap data in it into a marker dosage and map list}
\usage{
phasedHapMap2mat(hm)
}
\arguments{
\item{hm}{The data.frame that has HapMap data in it}
}
\value{
list with markers and map objects
}
\description{
Transform a data.frame with a hapmap data in it into a marker dosage and map list
}

