% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictValue.R
\name{predictValue}
\alias{predictValue}
\title{Genomic prediction}
\usage{
predictValue(sEnv = simEnv, popID = NULL, trainingPopID = NULL,
  locations = NULL, years = NULL, sharingInfo = "none")
}
\arguments{
\item{sEnv}{the environment that BSL functions operate in. Default is "simEnv" so use that to avoid specifying when calling functions}

\item{popID}{population ID to be predicted (default: the latest population)}

\item{trainingPopID}{population ID to be used for training a prediction model (default: all populations with phenotype data). NOTE: if sharingInfo="none" this parameter has no effect.}

\item{locations}{data from which locations should be used (default: all locations)}

\item{years}{data from which years should be used (default: all years)}

\item{sharingInfo}{one of "none", "markers", "pedigree".  If none, genotypic values are assumed IID. If markers or pedigree, a genomic or pedigree relationship matrix is constructed}
}
\value{
predicted values and the all information created before (list)
}
\description{
Genomic prediction
}

