% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfFertilize.R
\name{selfFertilize}
\alias{selfFertilize}
\title{Self-fertilize}
\usage{
selfFertilize(sEnv = simEnv, nProgeny = 100, popID = NULL)
}
\arguments{
\item{sEnv}{the environment that BSL functions operate in. Default is "simEnv" so use that to avoid specifying when calling functions}

\item{nProgeny}{the number of progeny}

\item{popID}{population ID to be self-fertilized (default: the latest population)}
}
\value{
sequence information of progenies and the all information created before (list)
}
\description{
Self-fertilize
}

